refactor(backend): isolate Tooling version policy
